requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.8
apex==0.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
backcall==0.2.0
bitarray==2.5.1
blis==0.7.9
boto3==1.26.20
botocore==1.29.20
cachetools==5.2.0
catalogue==2.0.8
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.5
confection==0.0.3
cycler==0.11.0
cymem==2.0.7
Cython==0.29.30
datasets==2.3.2
debugpy==1.6.3
decorator==5.1.1
dill==0.3.5.1
docker-pycreds==0.4.0
einops==0.4.1
en-core-web-sm @ file:///data10T/huangshijue/CMGCN-main/en_core_web_sm-3.4.1.tar.gz
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.4.1/en_core_web_trf-3.4.1-py3-none-any.whl
entrypoints==0.4
filelock==3.7.1
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
ftfy==6.0.3
future==0.18.2
fvcore==0.1.5.post20221122
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
h5py==2.10.0
huggingface-hub==0.10.1
hydra-core==1.0.7
idna==3.3
imageio==2.19.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
iopath==0.1.10
ipykernel==6.16.2
ipython==7.34.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jupyter_client==7.4.7
jupyter_core==4.11.2
kiwisolver==1.4.3
langcodes==3.3.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mock==4.0.3
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.9
nest-asyncio==1.5.6
networkx==2.6.3
nltk==3.7
numpy==1.21.6
oauthlib==3.2.0
omegaconf==2.0.6
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathtools==0.1.2
pathy==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
portalocker==2.4.0
preshed==3.0.8
promise==2.3
prompt-toolkit==3.0.32
protobuf==3.19.4
psutil==5.9.2
ptyprocess==0.7.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocoevalcap==1.2
pycocotools==2.0.4
pycparser==2.21
pydantic==1.10.2
pyDeprecate==0.3.2
pydot==1.4.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-crf==0.7.2
pytorch-lightning==1.6.4
pytorch-pretrained-bert==0.6.2
pytorch-pretrained-vit==0.0.7
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==24.0.1
regex==2022.6.2
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rouge-score==0.0.4
rsa==4.8
s3transfer==0.6.0
sacrebleu==2.1.0
sacremoses==0.0.53
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.9.8
seqeval==1.2.2
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.4.3
spacy-alignments==0.8.6
spacy-legacy==3.0.10
spacy-loggers==1.0.3
spacy-transformers==1.1.8
srsly==2.4.5
tabulate==0.8.10
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4.1
termcolor==2.1.1
thinc==8.1.5
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.5.4
tokenizers==0.12.1
torch==1.12.0
torch-cluster==1.6.0+pt112cu102
torch-geometric==2.1.0.post1
torch-scatter==2.1.0+pt112cu102
torch-sparse==0.6.15+pt112cu102
torch-spline-conv==1.2.1+pt112cu102
torchmetrics==0.9.2
torchvision==0.13.0
tornado==6.2
tqdm==4.64.0
traitlets==5.5.0
transformers==4.23.1
typer==0.7.0
typing_extensions==4.1.1
urllib3==1.26.12
visualize==0.5.0
wandb==0.13.3
wasabi==0.10.1
wcwidth==0.2.5
Werkzeug==2.1.2
wordninja==2.0.0
xxhash==3.0.0
yacs==0.1.8
yarl==1.7.2
zipp==3.8.0
